without npm

+++ src/filecoin-api-examples/DealRewarder.sol
@@ -1,15 +1,15 @@
 // SPDX-License-Identifier: MIT
 pragma solidity ^0.8.13;
 
-import { MarketAPI } from "@zondax/filecoin-solidity/contracts/v0.8/MarketAPI.sol";
-import { CommonTypes } from "@zondax/filecoin-solidity/contracts/v0.8/types/CommonTypes.sol";
-import { MarketTypes } from "@zondax/filecoin-solidity/contracts/v0.8/types/MarketTypes.sol";
-import { Actor } from "@zondax/filecoin-solidity/contracts/v0.8/utils/Actor.sol";
-import { Misc } from "@zondax/filecoin-solidity/contracts/v0.8/utils/Misc.sol";
+import {MarketAPI} from "@filecoin/contracts/v0.8/MarketAPI.sol";
+import {CommonTypes} from "@filecoin/contracts/v0.8/types/CommonTypes.sol";
+import {MarketTypes} from "@filecoin/contracts/v0.8/types/MarketTypes.sol";
+import {Actor} from "@filecoin/contracts/v0.8/utils/Actor.sol";
+import {Misc} from "@filecoin/contracts/v0.8/utils/Misc.sol";
 
-/* 
+/*
 Contract Usage
-    Step   |   Who   |    What is happening  |   Why 
+    Step   |   Who   |    What is happening  |   Why
     ------------------------------------------------
     Deploy | contract owner   | contract owner deploys address is owner who can call addCID  | create contract setting up rules to follow
     AddCID | data pinners     | set up cids that the contract will incentivize in deals      | add request for a deal in the filecoin network, "store data" function
@@ -19,7 +19,7 @@ Contract Usage
 */
 contract DealRewarder {
     mapping(bytes => bool) public cidSet;
-    mapping(bytes => uint) public cidSizes;
+    mapping(bytes => uint256) public cidSizes;
     mapping(bytes => mapping(uint64 => bool)) public cidProviders;
 
     address public owner;
@@ -33,10 +33,10 @@ contract DealRewarder {
 
     function fund(uint64 unused) public payable {}
 
-    function addCID(bytes calldata cidraw, uint size) public {
-       require(msg.sender == owner);
-       cidSet[cidraw] = true;
-       cidSizes[cidraw] = size;
+    function addCID(bytes calldata cidraw, uint256 size) public {
+        require(msg.sender == owner);
+        cidSet[cidraw] = true;
+        cidSizes[cidraw] = size;
     }
 
     function policyOK(bytes memory cidraw, uint64 provider) internal view returns (bool) {
@@ -44,33 +44,38 @@ contract DealRewarder {
         return !alreadyStoring;
     }
 
-    function authorizeData(bytes memory cidraw, uint64 provider, uint size) public {
+    function authorizeData(bytes memory cidraw, uint64 provider, uint256 size) public {
         require(cidSet[cidraw], "cid must be added before authorizing");
         require(cidSizes[cidraw] == size, "data size must match expected");
         require(policyOK(cidraw, provider), "deal failed policy check: has provider already claimed this cid?");
 
         cidProviders[cidraw][provider] = true;
     }
+
     type FilActorId is uint64;
+
     function claim_bounty(uint64 deal_id) public {
         MarketTypes.GetDealDataCommitmentReturn memory commitmentRet = MarketAPI.getDealDataCommitment(deal_id);
         uint64 providerRet = MarketAPI.getDealProvider(deal_id);
 
         authorizeData(commitmentRet.data, providerRet, commitmentRet.size);
-        
+
         // get dealer (bounty hunter client)
         uint64 clientRet = MarketAPI.getDealClient(deal_id);
 
-        // send reward to client 
+        // send reward to client
         send(clientRet);
 
-        // send reward to client 
+        // send reward to client
         send(clientRet);
-
     }
 
-    function call_actor_id(uint64 method, uint256 value, uint64 flags, uint64 codec, bytes memory params, uint64 id) public returns (bool, int256, uint64, bytes memory) {
-        (bool success, bytes memory data) = address(CALL_ACTOR_ID).delegatecall(abi.encode(method, value, flags, codec, params, id));
+    function call_actor_id(uint64 method, uint256 value, uint64 flags, uint64 codec, bytes memory params, uint64 id)
+        public
+        returns (bool, int256, uint64, bytes memory)
+    {
+        (bool success, bytes memory data) =
+            address(CALL_ACTOR_ID).delegatecall(abi.encode(method, value, flags, codec, params, id));
         (int256 exit, uint64 return_codec, bytes memory return_value) = abi.decode(data, (int256, uint64, bytes));
         return (success, exit, return_codec, return_value);
     }
@@ -80,8 +85,7 @@ contract DealRewarder {
         bytes memory emptyParams = "";
         delete emptyParams;
 
-        uint oneFIL = 1000000000000000000;
+        uint256 oneFIL = 1000000000000000000;
         Actor.callByID(CommonTypes.FilActorId.wrap(actorID), METHOD_SEND, Misc.NONE_CODEC, emptyParams, oneFIL, false);
     }
 }
-
